withher/matching-service/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
numpy==1.26.4
scipy==1.14.0
psycopg2-binary==2.9.9
sentence-transformers==3.0.1
python-dotenv==1.0.1
